Shifted genre functionality, exclusion genre now workin
Genre dropdown now works by passing options to the component rather than getting the options from within the component itself to avoid problems with both using the same state logic and one not populating with results. Also, exclusion now works as intended, however there is still some things that are going to need to be worked out. For example, some search queries on exclude seem to enver actually finish running, and i need to figure out what that problem is.