Started work on restructuring genre search
Instead of relying on getting every single manga through jikan, and then picking randomly from that, now we find the last available page with the given genre selections, choose a random page based off of that number, and then choose a random manga from that page. This is a better way to solve this issue as it gets around the jikan api rate limits, meaning that there are no longer going to be searches that last upwards of 10+ minutes. There are still kinks to be worked out from this, cause it is not perfect as images and stuff are not showing up on the pag itself, but this is a great start