Added loading bars and animations to pages
Through react-spinners package, I have changed it so instead of random static text, you now get a loading bar that shows when a search is taking place, and have made it so the loading takes place within the same card that holds the manga image and synopsis. This is in tandem with the cards fading in an out to look more smooth compared to the immediate jumping in and out of the cards that I had before hand. There are still multiple things to work on, such as thinking of better ways to handle how the cards sit on the page (i.e. if 5they should start as blank and then fade in on first search or always having some sort of card on screen), and if more information needs to be present on the screen when returning information such as a link to follow. This is on top of other general visual improvements that still need to be made.