Started implementation of external links
Adding the ability for the link of a manga on some other site to be made available to the user when they get a title. This is to make it easier for them to find a title on a website instead of needing to go searching manually. Still have more work to do, especially on the visual side of things, but the backend logic for the mangadex side of things is now functional. All credit to them for providing an easy to use API. Still need to figure out best way to provide them credit as well, need to research more into that later