Buttons and redirects, refactor of Jikan route
Created button for mangadex that, when clicked, takes you to the mangadex page of the title that came up (provided that the title is there). Also includes little note about the fuzzy/partial matching of the titles and why this can lead to incorrect page redirects.

Also reworked main random manga search to exclude genres that are not suitable for all audiences to make the website a bit more safe for the average user. There is still the potential to find results that are still NSFW, but if i go any further in restrictions many titles that are fine would be removed.

Finally, created button for battwo redirect. This does nothing but redirect to battwo api currently, and is essentially a just a placeholder, but will be used in the future