Mangadex button functional with warning/animation
Animated button fades in and out depending on search results. Fades in mangadex button if a valid mangadex entry is retrieved, and fades out when fetching a new manga. Also fades in an error message if a valid entry cannot be found telling users to seek elsewhere. After this, still need to have a popup to tell why this happens, and then need to translate all this stuff over to genre search