Some starting styling added to header, fixed error message issue
Added some basic styling for the header of the website. However, this is just basic boiler plating and more of a template on how to continue in the future than a final product. Still don't entirely know if i like the underline effect enough, or if i prefered the highlight boxes that i had before hand.   I also want to work on other ways i can add more visual appeal to it, cause as of now it still looks kinda bland.

Error message fix was a simple removal of a line that made a red text error appear when a title could not be found on mangadex in the middle of the page, that looked super ugly. This is just on the page without the genre filtering, so i might have to make the same change there in the future. Just dont have the time right now.