Major Styling Changes
Major changes to the colors and styling of the pages of the website to be more in line with what I personally enjoy, and what I think looks the best. I also changed the warning messages to be a component on their own, that i can reuse on each of the different pages that require it. Still more work to be done, but for now this is a good step. Another thing to look to doing is creating a proper error message when a search with the genre search is too specific and no results are returned/how to handle how manga results are shown (if website should display old results or get rid of results entirely)