Some more small changes
Changes to styling and text on some components. Started setup for making titles of page dynamic. Going to try and figure out a way to have the title dynamically updated based on the result that is returned, like how the title of the tab changes when clicking on a youtube video. Also need to find some sort of image to use as an svg (probably something Hat Kid related cause theme of the channel)