Metadata changes
File structure of different webpages had to be changed for metadata changes to be possible. Metadata import is not able to be used on client components, something that most of the pages were. Had to change file structure to make actual page structure a component to keep each page.tsx a server component and then importing the client component. Main page wasn't changed as required metadata was already default for the main page. Still need to think of way of implementing dynamic generateMetadata based on what results are returned