added remove warning feature
Can now remove the warning from the page if that is something that is required. Also added hover of the x to close the warning so it is apparent that is what is being clicked. can still add more to this, such as making it a persistent deletion, or making it so when you click the x, the warning just gets minimized to a page that can be reopened without needing to reload the page