Updated Markdown
Created full markdown page that includes things such as website features, proper acknowledgements to where I am getting information/api resources, and instruction on how to run the website locally. More can be done on this in the future, but I feel like this is a good bookend for this project as of now.