Update to include read me on main page of repo
Read me file was inside one of the folders in the repo. so copied that and made whatever necessary changes needed in order to include read me on the main page of the repo as well, so it can be readily apparent as to what is going on in the repo immediately upon accessing it instead of first needing to click into another page.